starts reorganization of folders to accommodate EMSphiInx
The Spherical Indexing routines will have a separate license/copyright statement, so we need to separate them out into their own library and source folders.
The EMsoftSphInxLib folder has sub-folders that reflect the original organization of Will's source files.  Both f90 and cpp files will be present in the same folders, and the CMake scripts will take care of compiling the new library and the programs.