reorganizes pyEMsoft building; adds pyEMsoftHDF wrapper
The .so files are now built in the regular EMsoft Build folder in a pyEMsoft subfolder; eventually the library files will be moved into the Bin folder and the user will need to set a proper path to find the pyEMsoft.py and pyEMsoftHDF.py modules.
This assumes that the user has properly set up the .bash_profile file using the "conda init bash" command.