modifies pattern binning in EMEBSDDI program
binning is now only available for the experimental patterns. The user must then correct the pattern center coordinates and detector pixel size to reflect the binning factor. The actual binning is done with the rescaler method from the ImageOPs module.  This is now tagged as version 5.0.3 once it is merged onto the develop branch